+++ src/sort.cpp
@@ -173,7 +173,7 @@ bool MergeMeshGLP(MeshGLP<Precision, I>& mesh) {
   sequence(vertNew2Old.begin(), vertNew2Old.end());
 
   radix_sort_with_key(vertNew2Old.begin(), vertNew2Old.end(),
-                      [&vertMorton](size_t i) { return vertMorton[i]; });
+                      [&vertMorton](int i) { return vertMorton[i]; });
 
   Permute(vertMorton, vertNew2Old);
   Permute(vertBox, vertNew2Old);
@@ -291,7 +291,7 @@ void Manifold::Impl::SortVerts() {
   sequence(vertNew2Old.begin(), vertNew2Old.end());
 
   radix_sort_with_key(vertNew2Old.begin(), vertNew2Old.end(),
-                      [&vertMorton](size_t i) { return vertMorton[i]; });
+                      [&vertMorton](int i) { return vertMorton[i]; });
 
   ReindexVerts(vertNew2Old, numVert);
 
@@ -404,7 +404,7 @@ void Manifold::Impl::SortFaces(Vec<Box>& faceBox, Vec<uint32_t>& faceMorton) {
   sequence(faceNew2Old.begin(), faceNew2Old.end());
 
   radix_sort_with_key(faceNew2Old.begin(), faceNew2Old.end(),
-                      [&faceMorton](size_t i) { return faceMorton[i]; });
+                      [&faceMorton](int i) { return faceMorton[i]; });
 
   // Tris were flagged for removal with pairedHalfedge = -1 and assigned kNoCode
   // to sort them to the end, which allows them to be removed.
